remove echo test from host open procedure
This is described in K1EL documentation someplace, but I must
have been rushed, because the docs suggest doing this before
sending the host open command, not after.  This didn't cause
problems with a WinKeyer USB (version code 30), but one user
with a K3NG based nanokeyer found this test (implemented
in the wrong part of the sequence) failed.

Removed the echo test, because the version returned by host_open
alone should be enough to validate an actual winkeyer device
is on that port.